+++ src/Form.php
@@ -332,6 +332,7 @@
      * @return \Fiv\Form\Element\Input
      */
     public function hidden($name, $value = null) {
+      trigger_error('Deprecated. use ' . Element\Hidden::class . ' class', E_USER_WARNING);
       $hidden = new  \Fiv\Form\Element\Input();
       $hidden->setType('hidden');
       $hidden->setName($name);